+++ notus/src/file_ops.rs
@@ -60,9 +60,9 @@ impl FilePair {
             } else {
                 let key_dir_entry = KeyDirEntry::new(
                     self.file_id,
-                    hint_entry.key_size() as u32,
-                    hint_entry.value_size() as u32,
-                    hint_entry.data_entry_position() as u32,
+                    hint_entry.key_size(),
+                    hint_entry.value_size(),
+                    hint_entry.data_entry_position(),
                 );
                 let raw_key: RawKey = bincode::deserialize(&hint_entry.key())?;
                 keys_dir.insert(raw_key.0, raw_key.1, key_dir_entry);
@@ -157,9 +157,9 @@ impl ActiveFilePair {
 
         Ok(KeyDirEntry::new(
             self.file_pair.file_id,
-            hint_entry.key_size() as u32,
-            hint_entry.value_size() as u32,
-            data_entry_position as u32,
+            hint_entry.key_size(),
+            hint_entry.value_size(),
+            data_entry_position,
         ))
     }
